feat: Support HTML output, and therefore EPUB via Pandoc
Readers reported that the PDF version is not usable on phones: and many
people these days only have phones. So some people attempted to convert
the output to EPUB themselves, using online tools, which produced
somewhat malformed output.

Instead of that, to make the book more accessible, we can use Typst's
in-development support for HTML output, and then convert that HTML to
EPUB using Pandoc.

The resultant EPUB is not particularly well-formatted at the moment,
because I do not yet know how to do that, but it is at least readable.